requirements.txt
streamlit>=1.25.0
matplotlib>=3.7.2
networkx>=3.1
openai>=0.28.0
plotly>=5.15.0
pygraphviz>=1.10
python-dotenv>=1.0.0
wordcloud>=1.9.2
twilio>=7.17.0
pandas>=2.0.3
numpy>=1.24.3